Upgrade to TensorFlow 2.15 (x86_64 + aarch64/GH200)
TF 2.15 is the ceiling that still ships tf.estimator + legacy Keras
optimizers (both removed in 2.16) and the lowest-risk window that
supports Hopper (sm_90) GPUs. Changes to run cleanly under it:

- optimizers.get(): request the legacy Keras optimizer via deserialize(
  use_legacy_optimizer=True) where the API exists (TF 2.11-2.15), since
  PiNN trains in estimator graph mode (tf.gradients + apply_gradients)
  which the new-style optimizer no longer supports.
- calculator: disable tf.data autotune/prefetch in the predict input_fn
  so the cached predictor reads the freshly-mutated atoms each call
  (TF>=2.15 prefetched ahead and returned the previous step's energy).
- io/tfr: read the dataset spec from the public dataset.element_spec
  instead of the private DatasetSpec._serialize() (fragile across the
  6-minor-version jump).
- io/ani: hoist `import h5py/numpy` to module scope (the generator ran
  outside load_ani's local imports -> NameError) and replace the
  h5py<3 Dataset.value with [()] (h5py>=3 ships with TF 2.15).
- utils.TuneTrainable: tf.train.CheckpointSaverListener ->
  tf.estimator.CheckpointSaverListener and tf.logging -> tf.get_logger()
  (both removed in the TF2 API).
- numpy>=1.24 removed-alias fixes: np.int -> int (cli), np.float -> float
  (io/cp2k). Repo-wide scan clean (only sized np.int32/64, np.float32).
- setup.py: TF extras bumped to >=2.15,<2.16, python_requires >=3.9,<3.12
  (drop EOL 3.8), protobuf<3.20 pin removed. CI: drop the 3.8 leg, add an
  experimental TF 2.15 / Py 3.10 leg.

BPNN jacobian tests (test_bpnn.py): the old test_jacob_bpnn compared two
separately-initialised models and relied on tf.random.set_seed making
their weights identical -- no longer true in TF>=2.15, so it failed.
Replace with weight-independent mathematical checks of the SF jacobian
dfp/dR:
- test_jacob_bpnn: jacobian vs full-autodiff forces with shared weights
  (set_weights) -- the chain-rule identity, tight tolerance.
- test_jacob_force_finite_difference: analytic -dE/dR vs central FD of E.
- test_jacob_matrix_finite_difference: dfp/dR via a fixed (untrainable)
  fingerprint read-out, checked against autodiff and central FD.
Full suite: 42 passed on x86 (TF 2.15.1).

Container + Nextflow plumbing for GPU clusters:
- Dockerfile -> TF 2.15.0; Dockerfile.gpu -> NGC tensorflow:24.03-tf2-py3
  (the only aarch64 GPU TF build); Singularity/Singularity.gpu mirrors.
- nextflow.config: `arrhenius` profile trains on the GPU partition and
  routes data prep (gen_*/get_qm7b) to the internet-enabled CPU
  partition, both via apptainer. Cluster-specific values are params with
  neutral defaults -- accounts (-A), gres, SIF paths, walltimes -- so no
  site account or personal path is baked into the config. train_time
  defaults to 2d to stay under a GrpTRESRunMins(gres/gpu) cap.
- datasets.nf: fetch rmd17 / mp2021 figshare files via the API with a
  browser UA (the article-zip endpoint returns 202 and rejects the
  default UA -> BadZipFile).

+++ pinn/cli.py
@@ -177,7 +177,7 @@ def log(logdir, tag, fmt):
     steps = [logs[k][:,0] for k in keys]
     data = [logs[k][:,1] for k in keys]
     steps, rows = np.unique(np.concatenate(steps), return_inverse=True)
-    cols = np.concatenate([np.full_like(v, i, np.int) for i, v in enumerate(data)])
+    cols = np.concatenate([np.full_like(v, i, int) for i, v in enumerate(data)])
 
     tmp = np.full([len(steps), len(logs.values())], np.nan)
     tmp[rows, cols] = np.concatenate(data)
